+++ plom/client/tools/rubric.py
@@ -24,14 +24,14 @@ class CommandGroupDeltaText(CommandTool):
     Note: must change mark
     """
 
-    def __init__(self, scene, pt, rid, meta, delta, text):
+    def __init__(self, scene, pt, rid, kind, delta, text):
         super().__init__(scene)
         self.gdt = GroupDeltaTextItem(
             pt,
             delta,
             text,
             rid,
-            meta,
+            kind,
             scene=scene,
             style=scene.style,
             fontsize=scene.fontSize,
@@ -78,12 +78,12 @@ class GroupDeltaTextItem(QGraphicsItemGroup):
     someone about building LaTeX... can we refactor that somehow?
     """
 
-    def __init__(self, pt, delta, text, rid, meta, scene, style, fontsize):
+    def __init__(self, pt, delta, text, rid, kind, scene, style, fontsize):
         super().__init__()
         self.pt = pt
         self.style = style
         self.rubricID = rid
-        self.meta = meta
+        self.kind = kind
         # centre under click
         self.di = DeltaItem(pt, delta, style=style, fontsize=fontsize)
         self.blurb = TextItem(
@@ -115,10 +115,6 @@ class GroupDeltaTextItem(QGraphicsItemGroup):
         else:
             self.blurb.setVisible(True)
             self.addToGroup(self.blurb)
-
-        # set up animators for delete
-        self.animator = [self.di, self.blurb]
-        self.animateFlag = False
 
         self.setFlag(QGraphicsItem.ItemIsMovable)
         self.setFlag(QGraphicsItem.ItemSendsGeometryChanges)
@@ -162,7 +158,7 @@ class GroupDeltaTextItem(QGraphicsItemGroup):
             self.pt.x() + self.x(),
             self.pt.y() + self.y(),
             self.rubricID,
-            self.meta,
+            self.kind,
             self.di.delta,
             self.blurb.getContents(),
         ]
@@ -212,7 +208,7 @@ class GhostComment(QGraphicsItemGroup):
         super().__init__()
         self.di = GhostDelta(dlt, fontsize)
         self.rubricID = "987654"  # a dummy value
-        self.meta = "relative"  # another dummy value
+        self.kind = "relative"  # another dummy value
         self.blurb = GhostText(txt, fontsize)
         self.changeComment(dlt, txt)
         self.setFlag(QGraphicsItem.ItemIsMovable)